Removed faulty test

+++ app/src/main/java/com/github/emilehreich/bootcamp/BoredActivity.kt
@@ -1,6 +1,8 @@
 package com.github.emilehreich.bootcamp
 
-data class BoredActivity(val activity: String,
-                         val type: String,
-                         val participants: Int,
-                         val price: Float)
+data class BoredActivity(
+    val activity: String,
+    val type: String,
+    val participants: Int,
+    val price: Double
+)
